♻️  refactor: Move CommandTrees to their own Containers + Plugins
