+++ rats-devtools/src/python/rats/docs/_plugin.py
@@ -6,7 +6,6 @@ import click
 # https://github.com/microsoft/pyright/issues/2882
 from rats import apps as apps
 from rats import cli as cli
-from rats import command_tree as command_tree
 from rats import devtools as devtools
 
 from ._commands import RatsCiCommands
@@ -18,23 +17,6 @@ class RatsDocsPlugin(apps.AnnotatedContainer):
 
     def __init__(self, app: apps.Container) -> None:
         self._app = app
-
-    @apps.group(command_tree.CommandTreeServices.GROUPS.subcommands("rats-devtools"))
-    def docs_command_tree(self) -> command_tree.CommandTree:
-        return command_tree.CommandTree(
-            name="docs",
-            description="Commands for managing documentation.",
-            children=tuple(
-                (
-                    child.to_command_tree(self)
-                    if isinstance(child, command_tree.CommandServiceTree)
-                    else child
-                )
-                for child in self._app.get_group(
-                    command_tree.CommandTreeServices.GROUPS.subcommands("rats-devtools docs")
-                )
-            ),
-        )
 
     @apps.group(devtools.AppServices.GROUPS.CLI_ROOT_COMMANDS)
     def docs_command(self) -> click.Command:
